requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
talos-api
sqlparse==0.3.1
# celery
# dogpile.cache
# redis
# gunicorn
# psycopg2-binary/pymysql
